Solutions/GabrielViana/requirements.txt
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
boto3==1.38.26
botocore==1.38.26
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
flask-restx==1.3.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
moto==5.1.5
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
responses==0.25.7
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
xmltodict==0.14.2
